Many of Change
There was a lot of changes between this update and the last, which is my fault I forgot to push a commit earlier. So far the app is in a fully functional state, with a view formatting issues that still need to be tackled. Will probably use for a little while to see what other changes or functions I would like to see within the app.