DailyTotals Bugs
There was an issue with the Daily Totals list not keeping an active count on the daily calories and protein, that has now been fixed. There was also an issue where it would delete daily data instead of saving it every night at 11:59pm, that is assumed to be fixed, if it is not you will be seeing another change log related to this tomorrow.