DailyTotals Fix & Main Spinner Rework
DailyTotals immediately started buggin so we went ahead and made a final fix for that (hopefully). The spinner on the main page has now been reworked as well so that if you stay on the main page, and are making multiple different entries, it will stay at whatever you are logging, i.e. many different foods for lunch. Do note that the spinner does reset if you re-open the app or go to another page within the app.