Re-build 1.1
The app is very broken in this state but does build. The database is currently broken, doesn't initially save or save meal logs on app close. With that being said everything else seems semi-functional, and this is the initial re-build of the app to be overall more efficient and allow for much more functionality within the app in the future. Plans on the inital horizon are adding a protein calculator, adding training data section, weight tracking section, meal prep section, tracking individual macros, with more ideas to come. The first stable version will be 1.1.1, although that version will most likely not have all of these features listed.